+++ src/Simulation/Sim_Types.cpp
@@ -58,8 +58,8 @@ sycl::range<1> get_compute_range(sycl::queue &q, uint32_t N_sims)
 {
     auto device = q.get_device();
     auto max_wg_size = device.get_info<sycl::info::device::max_work_group_size>();
-    double d_sims = N_sims;
-    double d_max_wg_size = max_wg_size;
+    float d_sims = N_sims;
+    float d_max_wg_size = max_wg_size;
     auto N_compute_units = static_cast<uint32_t>(std::ceil(d_sims / d_max_wg_size));
     return N_compute_units;
 }
